Block Editor: Move 'ParentSelectorMenuItem' into a separate folder
